Move menu classes into a separate file
